WIP: build a lit-element v2.0.0
