+++ generators/app/index.js
@@ -102,7 +102,7 @@ module.exports = class extends Generator {
     this.fs.copyTpl(`${this.templatePath()}/*`, this.destinationPath(), this.props);
 
     this.fs.copyTpl(
-      `${this.templatePath()}/**/.!(gitignore)*`,
+      `${this.templatePath()}/**/.!(gitignorefile)*`,
       this.destinationRoot(),
       this.props
     );
@@ -121,13 +121,19 @@ module.exports = class extends Generator {
 
     this.fs.copyTpl(
       this.templatePath(`src/_element.js`),
-      this.destinationPath(`${elementName}.js`),
+      this.destinationPath(`src/${elementName}.js`),
       this.props
     );
 
     this.fs.copyTpl(
       this.templatePath(`src/_element-styles.js`),
-      this.destinationPath(`${elementName}-styles.js`),
+      this.destinationPath(`src/${elementName}-styles.js`),
+      this.props
+    );
+
+    this.fs.copyTpl(
+      this.templatePath(`src/_element-base.js`),
+      this.destinationPath(`src/${elementName}-base.js`),
       this.props
     );
   }
